Moved 'fortune cookie' functionality into module.
