+++ meadowlark.js
@@ -12,13 +12,7 @@ app.set('port', process.env.PORT || 3000);
 
 app.use(express.static(__dirname + '/public'));
 
-var fortuneCookies = [
-	"Conquer your fears or they will conquer you.",
-	"Rivers need springs.",
-	"Do not fear what you don't know.",
-	"You will have a pleasant surprise.",
-	"Whenever possible, keep it simple.",
-];
+var fortune = require('./lib/fortune.js');
 
 app.get('/', function(req, res) {
 	res.render('home');
@@ -26,7 +20,7 @@ app.get('/', function(req, res) {
 app.get('/about', function(req,res){
 	var randomFortune = 
 		fortuneCookies[Math.floor(Math.random() * fortuneCookies.length)];
-	res.render('about', { fortune: randomFortune });
+	res.render('about', { fortune: fortune.getFortune() });
 });
 
 // 404 catch-all handler (middleware)
